change: undo AST `proto3-suite` seperation
