--- proto3-suite-ast/fourmolu.yaml
@@ -1,8 +1,0 @@
-indentation: 2
-comma-style: "leading"
-indent-wheres: true
-record-brace-space: true
-diff-friendly-import-export: false
-respectful: true
-newlines-between-decls: 1
-haddock-style: "single-line"
