+++ src/Proto3/Suite/DotProto/AST/Lens.hs
@@ -1,9 +1,10 @@
-{-# LANGUAGE TemplateHaskell            #-}
+{-# LANGUAGE TemplateHaskell #-}
+
+-- |
 module Proto3.Suite.DotProto.AST.Lens where
 
 import Control.Lens.TH
-import Proto3.Suite.DotProto.AST
+import Proto3.Suite.DotProto.AST.Message
 
 makePrisms ''DotProtoDefinition
 makePrisms ''DotProtoMessagePart
-
