+++ src/Proto3/Suite/DotProto/AST/Message.hs
@@ -1,3 +1,6 @@
+{-# LANGUAGE DerivingStrategies #-}
+{-# LANGUAGE DerivingVia #-}
+
 -- |
 --
 -- @since 1.0.0
@@ -41,10 +44,10 @@ where
 
 import Data.String (IsString)
 
-import Proto3.Suite.DotProto.AST.Core (DotProtoIdentifier)
-import Proto3.Suite.DotProto.AST.Option (DotProtoOption)
 import Proto3.Suite.DotProto.AST.Enumerate (DotProtoEnumPart)
 import Proto3.Suite.DotProto.AST.Field (DotProtoField, DotProtoReservedField)
+import Proto3.Suite.DotProto.AST.Identifier (DotProtoIdentifier)
+import Proto3.Suite.DotProto.AST.Option (DotProtoOption)
 
 --------------------------------------------------------------------------------
