+++ src/Proto3/Suite/DotProto/AST/Type.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE DerivingStrategies #-}
+
 -- |
 --
 -- @since 1.0.0
@@ -27,7 +29,8 @@ module Proto3.Suite.DotProto.AST.Type
   )
 where
 
-import Proto3.Suite.DotProto.AST.Core (DotProtoIdentifier)
+import Proto3.Suite.DotProto.AST.Identifier (DotProtoIdentifier)
+import Proto3.Suite.DotProto.AST.Value (DotProtoValue)
 
 --------------------------------------------------------------------------------
